+++ frontend/src/routes.js
@@ -1,0 +1,81 @@
+// import
+import Dashboard from "./screens/Dashboard/Dashboard";
+import Tables from "./screens/Dashboard/Tables";
+import Billing from "./screens/Dashboard/Billing";
+
+import Profile from "./screens/Dashboard/Profile";
+import SignIn from "./screens/LogInPage.js";
+import SignUp from "./screens/RegistrationPage.js";
+
+import {
+  HomeIcon,
+  StatsIcon,
+  CreditIcon,
+  PersonIcon,
+  DocumentIcon,
+  RocketIcon,
+  SupportIcon,
+} from "./components/components/Icons/Icons";
+
+var dashRoutes = [
+  {
+    path: "/dashboard",
+    name: "Dashboard",
+
+    icon: <HomeIcon color="inherit" />,
+    component: Dashboard,
+    layout: "/admin",
+  },
+  {
+    path: "/tables",
+    name: "Tables",
+
+    icon: <StatsIcon color="inherit" />,
+    component: Tables,
+    layout: "/admin",
+  },
+  {
+    path: "/billing",
+    name: "Billing",
+
+    icon: <CreditIcon color="inherit" />,
+    component: Billing,
+    layout: "/admin",
+  },
+
+  {
+    name: "ACCOUNT PAGES",
+    category: "account",
+
+    state: "pageCollapse",
+    views: [
+      {
+        path: "/profile",
+        name: "Profile",
+
+        icon: <PersonIcon color="inherit" />,
+        secondaryNavbar: true,
+        component: Profile,
+        layout: "/admin",
+      },
+      {
+        path: "/signin",
+        name: "Sign In",
+
+        icon: <DocumentIcon color="inherit" />,
+        component: SignIn,
+        layout: "/auth",
+      },
+      {
+        path: "/signup",
+        name: "Sign Up",
+
+        icon: <RocketIcon color="inherit" />,
+        secondaryNavbar: true,
+        component: SignUp,
+        layout: "/auth",
+      },
+    ],
+  },
+];
+export default dashRoutes;