+++ CHANGELOG.md
@@ -7,10 +7,13 @@
 ### Добавлено
 
 - Docstrings for new func in user service [@fringer2423](https://github.com/fringer2423)
+- Purity Ui Dashboard [@AlexTrubkina](https://github.com/AlexTrubkina)
 
 ### Изменено
 
 - User service [@fringer2423](https://github.com/fringer2423)
+- log-in button [@AlexTrubkina](https://github.com/AlexTrubkina)
+- blueBall.css
 
 ## [0.0.31] - 2022-11-17